Renamed tab options file
